Improvements of pbxplore module architecture. Scripts have been also modified to reflect these changes
