+++ pbxplore/__init__.py
@@ -1,4 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .formats.loader import chains_from_files, chains_from_trajectory
+from .structure.loader import chains_from_files, chains_from_trajectory
+from .assign import assign
+from . import PB
