+++ pbxplore/analysis/__init__.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .clustering import hclust
-from .neq import compute_neq, write_neq
+from .clustering import hclust, distance_matrix, RError
+from .count import count_matrix, read_occurence_file
+from .neq import compute_neq
+from .utils import substitution_score, compute_freq_matrix, compute_score_by_position
+from .visualization import generate_weblogo, plot_map, plot_neq
