+++ pbxplore/analysis/neq.py
@@ -10,7 +10,7 @@ import math
 import numpy
 
 # Local module
-from .. import utils
+from . import utils
 
 
 def _neq_per_residue(row_freq):
@@ -56,28 +56,3 @@ def compute_neq(count_mat):
     neq_array = numpy.apply_along_axis(_neq_per_residue, 1, freq)
 
     return neq_array
-
-
-def write_neq(outfile, neq_array, residue_min=1, residue_max=None):
-    """
-    Write the Neq matrix in an open file
-
-    Parameters
-    ----------
-    outfile : file descriptor
-        The file descriptor to write in. It must allow writing.
-    neq_array : numpy array
-        a 1D array containing the neq values.
-    residue_min: int
-        the lower bound of residue frame
-    residue_max: int
-        the upper bound of residue frame
-
-    """
-
-    # Slice
-    neq = utils._slice_matrix(neq_array, residue_min, residue_max)
-
-    print("%-6s %8s " % ("resid", "Neq"), file=outfile)
-    for (res, neq) in enumerate(neq):
-        print("%-6d %8.2f " % (res + residue_min, neq), file=outfile)
