+++ pbxplore/io/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .fasta import read_fasta, read_several_fasta, write_fasta
+from .write import write_phipsi, write_flat, write_count_matrix, write_neq
