+++ pbxplore/structure/__init__.py
@@ -2,3 +2,4 @@
 # -*- coding: utf-8 -*-
 
 from .PDB import PDB_EXTENSIONS, PDBx_EXTENSIONS
+from .loader import chains_from_files, chains_from_trajectory
